Split out spotipy wrapper code to its own file for clarity
